System clock initialization function defined.
System clock source is set to 8MHz HSE, which is supplied to PLL
engine to generate 168MHz clock speed.
Main system bus prescaler (AHB1 prescaler) is set to 1, so system
clock (HCLK) operates with the speed of 168MHz.
APB1 prescaler is set to 4, so all peripherals connected to APB1 bus
are supplied with 42MHz clock speed.
APB2 prescaler is set to 2, so all peripherals connected to APB2 bus
are supplied with 84MHz clock speed.
These speeds are maximum for target microcontroller.